fix(myorganizer): wrap searchParams pages in Suspense
